Move `DependencyGroups` to uv-pypi-types so it can be imported there
