+++ crates/uv-pypi-types/src/lib.rs
@@ -1,5 +1,6 @@
 pub use base_url::*;
 pub use conflicts::*;
+pub use dependency_groups::*;
 pub use direct_url::*;
 pub use lenient_requirement::*;
 pub use marker_environment::*;
@@ -12,6 +13,7 @@ pub use supported_environments::*;
 
 mod base_url;
 mod conflicts;
+mod dependency_groups;
 mod direct_url;
 mod lenient_requirement;
 mod marker_environment;
