+++ crates/uv-workspace/Cargo.toml
@@ -25,7 +25,7 @@ uv-normalize = { workspace = true }
 uv-options-metadata = { workspace = true }
 uv-pep440 = { workspace = true }
 uv-pep508 = { workspace = true }
-uv-pypi-types = { workspace = true }
+uv-pypi-types = { workspace = true, features = ["serde"] }
 uv-static = { workspace = true }
 uv-warnings = { workspace = true }
